+++ packages/plugins/next-intl/src/ideExtension/messageReferenceMatchers.ts
@@ -1,71 +1,41 @@
 /* eslint-disable @typescript-eslint/no-non-null-assertion */
-/**
- * Using parsimmon because:
- *
- * 1. Chevrotain is too complicated.
- * 2. TypeScript's compiler doesn't work in the browser.
- * 3. TypeScripts compiler
- */
-
 import Parsimmon from "parsimmon"
 import type { PluginSettings } from "../settings.js"
 
-const createParser = (settings: PluginSettings, defaultNS: string | undefined) => {
-	// Create a Parsimmon language
+const createParser = (settings: PluginSettings, namespaces: Record<number, string>) => {
 	return Parsimmon.createLanguage({
-		// The entry point for message reference matching.
-		//
-		// 1. Match a t function call or any other character.
-		// 2. Match as many of these as possible.
-		// 3. Filter out any non-object matches.
 		entry: (r) => {
 			return Parsimmon.alt(r.FunctionCall!, Parsimmon.any)
 				.many()
 				.map((matches) => {
-					// filter arbitrary characters
 					return matches.filter((match) => typeof match === "object")
 				})
 		},
 
-		// A string literal is either a single or double quoted string
 		stringLiteral: (r) => {
 			return Parsimmon.alt(r.doubleQuotedString!, r.singleQuotedString!)
 		},
 
-		// Double quoted string literal parser
-		//
-		// 1. Start with a double quote.
-		// 2. Then match any character that is not a double quote.
-		// 3. End with a double quote.
 		doubleQuotedString: () => {
 			return Parsimmon.string('"').then(Parsimmon.regex(/[^"]*/)).skip(Parsimmon.string('"'))
 		},
 
-		// Single quoted string literal parser
-		//
-		// 1. Start with a single quote.
-		// 2. Then match any character that is not a single quote.
-		// 3. End with a single quote.
 		singleQuotedString: () => {
 			return Parsimmon.string("'").then(Parsimmon.regex(/[^']*/)).skip(Parsimmon.string("'"))
 		},
 
-		// Whitespace parser
 		whitespace: () => {
 			return Parsimmon.optWhitespace
 		},
 
-		// Colon parser
 		colon: (r) => {
 			return Parsimmon.string(":").trim(r.whitespace!)
 		},
 
-		// Comma parser
 		comma: (r) => {
 			return Parsimmon.string(",").trim(r.whitespace!)
 		},
 
-		// Parser for t function calls
 		FunctionCall: function (r) {
 			return Parsimmon.seqMap(
 				Parsimmon.regex(/[^a-zA-Z0-9]/), // no preceding letters or numbers
@@ -77,9 +47,18 @@ const createParser = (settings: PluginSettings, defaultNS: string | undefined) =
 				Parsimmon.regex(/[^)]*/), // ignore the rest of the function call
 				Parsimmon.string(")"), // end with a closing parenthesis
 				(_, __, ___, start, messageId, end) => {
-					// -- handle namespaces --
-					if (defaultNS) {
-						messageId = defaultNS + "." + messageId
+					// Find the most recent namespace defined before this message id
+					const nearestNamespace = Object.keys(namespaces)
+						.map(Number)
+						.filter((index) => index <= start.offset)
+						.sort((a, b) => b - a)[0]
+
+					if (nearestNamespace) {
+						const namespace = namespaces[nearestNamespace]
+
+						if (namespace && nearestNamespace) {
+							messageId = `${namespace}.${messageId}`
+						}
 					}
 
 					return {
@@ -94,20 +73,19 @@ const createParser = (settings: PluginSettings, defaultNS: string | undefined) =
 								character: end.column,
 							},
 						},
-					} // satisfies MessageReferenceMatch
+					}
 				}
 			)
 		},
 	})
 }
 
-const createNamespaceParser = (pattern: string) => {
+const createNamespaceParser = () => {
 	return Parsimmon.createLanguage({
 		entry: (r) => {
 			return Parsimmon.alt(r.FunctionCall!, Parsimmon.any)
 				.many()
 				.map((matches) => {
-					// filter arbitrary characters
 					return matches.filter((match) => typeof match === "object")
 				})
 		},
@@ -136,80 +114,84 @@ const createNamespaceParser = (pattern: string) => {
 			return Parsimmon.string(":").trim(r.whitespace!)
 		},
 
-		nsValue: (r) => {
-			return Parsimmon.seq(
-				r.whitespace!,
-				Parsimmon.string("namespace").trim(r.whitespace!).skip(r.colon!), // namespace key parser
-				r.stringLiteral!.trim(r.whitespace!)
-			).map(([, , val]) => `${val}`)
+		stringParser: () => {
+			return Parsimmon.lookahead(Parsimmon.noneOf("{")).then(
+				Parsimmon.regexp(/"([^"]*)"/, 1).map((str) => ({ type: "string", value: str }))
+			)
 		},
 
-		// Parser for t function calls
+		objectParser: () => {
+			const x = {
+				stringLiteral: Parsimmon.regexp(/"([^"\\]|\\.)*"/).map((value) => JSON.parse(value)), // Match double-quoted strings and parse them
+				variableName: Parsimmon.regexp(/[a-zA-Z_][a-zA-Z0-9_]*(\.[a-zA-Z_][a-zA-Z0-9_]*)*/), // Match variable names
+				functionLiteral: Parsimmon.regexp(/function\s*\([^)]*\)\s*\{[^}]*\}/), // Match function literals
+			}
+
+			const valueParser = Parsimmon.alt(x.stringLiteral, x.variableName, x.functionLiteral)
+
+			const propertyParser = Parsimmon.seq(
+				Parsimmon.regexp(/[\w]+/).skip(Parsimmon.optWhitespace), // Property name
+				Parsimmon.string(":").then(Parsimmon.optWhitespace).then(valueParser).fallback(undefined) // Optional colon and value
+			).map(([key, value]) => (value !== null ? [key, value] : [key, undefined]))
+
+			return Parsimmon.optWhitespace.then(
+				Parsimmon.seq(
+					Parsimmon.string("{").then(Parsimmon.optWhitespace),
+					Parsimmon.sepBy(propertyParser, Parsimmon.string(",").then(Parsimmon.optWhitespace)),
+					Parsimmon.optWhitespace.then(Parsimmon.string("}"))
+				).map(([, entries]) => {
+					const obj = Object.fromEntries(entries)
+					return { type: "object", value: obj }
+				})
+			)
+		},
+
 		FunctionCall: function (r) {
 			return Parsimmon.seqMap(
 				Parsimmon.regex(/[^a-zA-Z0-9]/), // no preceding letters or numbers
-				Parsimmon.regex(/\b(?:use|get)Translations\b/), // starts with useTranslation or getTranslations
-				Parsimmon.string(pattern), // then an opening parenthesis
+				Parsimmon.regex(/\b(?:use|get)Translations\b/), // starts with useTranslations or getTranslations
+				Parsimmon.string("("), // then an opening parenthesis
 				Parsimmon.index, // start position of the message id
-				Parsimmon.regex(/[^)]*/), // inside string
+				Parsimmon.alt(r.stringParser!, r.objectParser!),
 				Parsimmon.index, // end position of the message id
 				Parsimmon.regex(/[^)]*/), // ignore the rest of the function call
 				Parsimmon.string(")"), // end with a closing parenthesis
-				(_, __, ___, ____, insideString) => {
-					const namespaceParser = Parsimmon.alt(
-						Parsimmon.seq(
-							Parsimmon.string("{").trim(r.whitespace!),
-							Parsimmon.string("locale").trim(r.whitespace!),
-							Parsimmon.alt(
-								Parsimmon.seq(r.comma!, r.nsValue!).map(([, val]) => val),
-								Parsimmon.seq(
-									r.colon!,
-									r.stringLiteral!.trim(r.whitespace!),
-									r.comma!,
-									r.nsValue!,
-									Parsimmon.regex(/[^)}]*/)
-								).map(([, , , val]) => val)
-							),
-							Parsimmon.string("}").trim(r.whitespace!)
-						).map(([, , val]) => val),
-						Parsimmon.seq(
-							Parsimmon.string("{").trim(r.whitespace!),
-							r.nsValue!,
-							Parsimmon.regex(/[^)}]*/),
-							Parsimmon.string("}").trim(r.whitespace!)
-						).map(([, val]) => val),
-						Parsimmon.seq(r.stringLiteral!)
-					)
-					const namespace = (
-						namespaceParser.parse(insideString) as { status: boolean; value: string }
-					).value
+				(_, __, ___, start, insideString, end) => {
+					let namespace = undefined
 
-					return {
-						ns: namespace,
+					if (insideString.type === "string") {
+						namespace = insideString.value
+					} else if (insideString && insideString.value.namespace) {
+						namespace = insideString.value.namespace
 					}
+					return { ns: namespace, start: start.offset, end: end.offset }
 				}
 			)
 		},
 	})
 }
 
-function parseNameSpaces(settings: PluginSettings, sourceCode: string) {
-	const namespaceParser = createNamespaceParser("(")
+function parseNameSpaces(sourceCode: string) {
+	const namespaceParser = createNamespaceParser()
 	const namespaces = namespaceParser.entry!.tryParse(sourceCode)
-	if (namespaces.length > 0) {
-		return namespaces[0].ns
-	} else {
-		return undefined
+	const namespaceMap: Record<number, string> = {}
+	for (const nsObj of namespaces) {
+		if (nsObj.ns) {
+			namespaceMap[nsObj.start] = nsObj.ns
+		}
 	}
+	return namespaceMap
 }
 
 // Parse the expression
 export function parse(sourceCode: string, settings: PluginSettings) {
 	try {
-		const namespace: undefined | string = parseNameSpaces(settings, sourceCode)
-		const parser = createParser(settings, namespace)
-		return parser.entry!.tryParse(sourceCode)
-	} catch {
+		const namespaces: Record<number, string> = parseNameSpaces(sourceCode)
+		const parser = createParser(settings, namespaces)
+		const result = parser.entry!.tryParse(sourceCode)
+		return result
+	} catch (error) {
+		console.error("Parsing error:", error)
 		return []
 	}
 }