+++ packages/plugins/json/CHANGELOG.md
@@ -1,5 +1,37 @@
 # @inlang/plugin-json
 
+## 5.1.48
+
+### Patch Changes
+
+- Updated dependencies [9d2aa1a]
+  - @inlang/sdk@0.35.6
+
+## 5.1.47
+
+### Patch Changes
+
+- Updated dependencies [64e30ee]
+  - @inlang/sdk@0.35.5
+
+## 5.1.46
+
+### Patch Changes
+
+- @inlang/sdk@0.35.4
+
+## 5.1.45
+
+### Patch Changes
+
+- @inlang/sdk@0.35.3
+
+## 5.1.44
+
+### Patch Changes
+
+- @inlang/sdk@0.35.2
+
 ## 5.1.43
 
 ### Patch Changes