+++ packages/versioned-interfaces/module/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @inlang/module
 
+## 1.2.13
+
+### Patch Changes
+
+- @inlang/plugin@2.4.13
+
 ## 1.2.12
 
 ### Patch Changes